glsl: add loop jump visitor
This visitor will allow us to determine all the loop jumps that
correspond to each loop. In SSA form, each input to a phi node is
associated with a predecessor basic block. In the case of phi nodes
at the beginning and end of loops, these predecessor blocks will
include all blocks that end with a loop_jump (break or continue), and
so in order to insert phi nodes we must know all the loop_jump
instructions that correspond to each loop.

v2:
-add more comments
-store outer_lj instead of looking it up (Paul Berry)

Reviewed-by: Paul Berry <stereotype441@gmail.com>