glsl: add SSA infrastructure
This patch introduces all the changes to the IR that are necessary for
representing programs in the SSA form. This consists of a new variable
mode, the SSA temporary, which is guarenteed to be written to exactly
once, and classes to represent phi nodes in the IR.

In the current code, variables are first declared using an ir_variable
instruction inserted into the instruction stream, and then every
dereference will point to the ir_variable declared earlier. SSA
temporaries, however, do not work this way. Instead, the variable
is declared when it is assigned. That is, the variable is "owned" by
the one and only instruction where it is defined.

In SSA, phi nodes may exist at the beginning of any "join nodes," or
basic blocks with more than one predecessor. In our IR, this can happen
in one of three places:

- After an if statement, where the then branch and the else branch
converge.
- At the beginning of a loop, which can be reached from either before
the loop (on the first iteration), the end of the loop (when we get to
the end of the loop and jump back to the beginning), or any continue
statement.
- At the end of a loop, which can be reached from any break statement
within the loop.

Accordingly, there are three different types of phi nodes: if phi nodes,
phi nodes at the beginning of a loop, and phi nodes at the end of a
loop, all of which derive from the ir_phi base class.

v2:

-make ir_phi constructor take the destination (Paul)
-merge ssa_phi, ssa_call, and ssa_assignment into ssa_owner (Paul)
-make ir_phi an abstract class and remove visit(ir_phi *) from ir_visitor
(Paul)
-fix bugs in ir_clone where ssa_owner sometimes wasn't being set
-simplify printing phi node sources (Paul)
-make the name of if/loop phi nodes more consistent when printing (Paul)
-add more invariants to ir_validate (Paul)