glsl: add a pass to convert out of SSA form
Right now we are being basically as naive as possible, and inserting
more copies than necessary. It is possible to implement a more
sophisticated algorithm later, although extending the current copy
propagation pass to support loops better and/or relying on backends to
do copy propagation may make this unecessary.

v2:
-fix potentially confusing comment (Paul)
-don't manually walk through function signatures (Paul)
-fix bug where ssa_owner wasn't getting set to NULL

Reviewed-by: Paul Berry <stereotype441@gmail.com>