i965/fs: be more careful about types when emitting textures
The code before worked when we set the correct type for inputs to
texture instructions, but with NIR we don't have type information so we
can, and do, load the coordinates with the "wrong" type - this shouldn't
matter, because for the most part all we're doing is moving bits around.
Make sure we do the right thing by correctly setting the types to the
moves and integer divide.