lima/gpir: Add a scheduling heuristic to value_regalloc
From experiments, it seems like the quality of the final schedule
heavily depends on which registers value_regalloc chooses. This makes
sense, since the initial schedule is very conservative, trying to
minimize the number of registers used. If we choose the wrong register
assignment, then we're locked into the conservative schedule thanks to
the false dependencies, and we can't reorder things to take less
instructions. While assigning registers round-robin helps, it still
isn't perfect. For example, if there's an expression that's the sum of
two large expressions:

output = expr1 + expr2

then the initial schedule will tend to put the nodes for one of the
expressions first, and then the other second, with no overlap:

r1 = <expr1> # many nodes...
r2 = <expr2> # many nodes...
output = add r1, r2

If it takes more than 11 nodes to compute expr1, it's guaranteed that
there will be a false dependency between nodes in expr1 and expr2, when
we might have been able to avoid it by assigning them to different
registers. At the very least, we should avoid giving a node inside expr
the same register as r1 gets, since then expr1 couldn't overlap with
that node or any of its children inside expr2, but this is guaranteed to
happen under the round-robin scheme if expr2 has more than 11 nodes.

To do something smarter, we need to know how our decisions are affecting
the final schedule. The critical path length ("dist" in the scheduler)
is a good proxy for this, since it's easy to compute and a good estimate
for how good the schedule will be. The goal of the value register
allocator will be to cause the minimum increase to dist. To do that, we
keep track of the dist of each node, and we keep enough information to
figure out how a given register assignment will increase dist due to the
write-after-read dependencies. The way the assignment is chosen is a
little subtle, as explained in the comments. But it seems to work well
in practice.

This improves a simple shader that just multiplies an attribute vec4 by
a uniform mat4 by 28%, making it as good as the blob compiler (although
the blob compiler also clamps 1/w, which we don't do yet, although I'm
not sure how much it matters). It also reduces the kmscube vertex shader
from 60 instructions to 47, only 6 more than the blob compiler.